+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * Add - Add support for Elementor Pro Page Builder Custom Header.
 * Add - Add support for Elementor Pro Page Builder Custom Footer.
 * Add - Add support for Elementor Pro Page Builder Custom 404 page.
+* Add - Add support for Elementor Pro Page Builder Custom single post.
+* Add - Add support for Elementor Pro Page Builder Custom archive page.
 
 #### 1.4.4 - May 03, 2018
 * Add - Add metabox option to hide or show breadcrumbs.